+++ methods/src/test/kotlin/methods/polynomials/PolynomialsRulesTest.kt
@@ -1,6 +1,7 @@
 package methods.polynomials
 
 import engine.methods.testRule
+import methods.polynomials.PolynomialRules.NormalizeMonomial
 import methods.polynomials.PolynomialRules.NormalizePolynomial
 import methods.polynomials.PolynomialRules.RearrangeProductOfMonomials
 import org.junit.jupiter.api.Test
@@ -30,5 +31,11 @@ class PolynomialsRulesTest {
         testRule("1 + 2y + [[y^2]/2]", NormalizePolynomial, "[[y^2]/2] + 2y + 1")
         testRule("[t^10] + 2[t^3] + sqrt[3] + 1", NormalizePolynomial, null)
         testRule("1 + [t^10] + 2[t^3] + sqrt[3]", NormalizePolynomial, "[t^10] + 2[t^3] + 1 + sqrt[3]")
+    }
+
+    @Test
+    fun testNormalizeMonomial() {
+        testRule("[x/2] + 1", NormalizeMonomial, "[1/2]x + 1")
+        testRule("[1/2]x - [3[x^2] / 2]", NormalizeMonomial, "[1/2]x - [3/2][x^2]")
     }
 }